requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.7.0
django-etc==1.3.0
django-filter==2.4.0
django-formtools==2.3
django-hitcount==1.3.3
django-markdown-deux==1.0.5
djangorestframework==3.12.4
future==0.18.2
idna==3.2
Markdown==3.3.4
markdown2==2.4.0
oauthlib==3.1.1
Pillow==8.3.1
protobuf==3.14.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
serializers==0.2.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.6
